--- webapp/src/home/home-page.js
@@ -1,30 +1,0 @@
-import React, { Fragment } from 'react'
-import { useQuery } from '@apollo/client'
-import GetTransactions from '../gql/transactions.gql'
-import { TxTable } from '../components/transactions/TxTable'
-
-export function Home () {
-  const { loading, error, data = {} } = useQuery(GetTransactions)
-
-  if (loading) {
-    return (
-      <Fragment>
-        Loading...
-      </Fragment>
-    )
-  }
-
-  if (error) {
-    return (
-      <Fragment>
-        ¯\_(ツ)_/¯
-      </Fragment>
-    )
-  }
-
-  return (
-    <Fragment>
-      <TxTable data={data.transactions} />
-    </Fragment>
-  )
-}
