+++ webapp/src/pages/home/home-page.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export function Home () {
+  return (
+    <div>
+      <h1>Welcome to your Page</h1>
+    </div>
+  )
+}
