+++ webapp/src/routes.js
@@ -1,25 +1,18 @@
 import React from 'react'
-import { BrowserRouter as Router, Route, Link } from 'react-router-dom'
+import { BrowserRouter as Router, Route } from 'react-router-dom'
 import { css } from '@emotion/core'
-import { Home } from './home'
+import { Home } from './pages/home'
+import NavBar from './components/nav/navbar'
+import Account from './pages/account/account-page'
 
 function AppRouter () {
   return (
     <Router>
       <div css={layoutStyle}>
-        <nav css={navStyle}>
-          <ul >
-            <li>
-              <Link to='/'>Home</Link>
-            </li>
-            <li>
-              <Link to='/another'>Another route</Link>
-            </li>
-          </ul>
-        </nav>
+        <NavBar />
         <div className='main-content' css={contentStyle}>
           <Route component={Home} exact path='/' />
-          <Route component={() => (<div>Content for /another route</div>)} exact path='/another' />
+          <Route component={Account} exact path='/Account' />
         </div>
       </div>
     </Router>
@@ -32,20 +25,6 @@ const layoutStyle = css`
     display: grid;
     grid-row-gap: 24px;
     padding: 8px;
-`
-
-const navStyle = css`
-  grid-row: 1;
-
-  & > ul {
-      display: flex;
-      flex-direction: row;
-      list-style-type: none;
-  }
-  
-  & > ul > li:not(:first-of-type) {
-    margin-left: 16px;
-  }
 `
 
 const contentStyle = css`
